+++ src/main/kotlin/de/unisaarland/cs/se/selab/plant/Cherry.kt
@@ -49,7 +49,7 @@ class Cherry : PlantationPlant() {
         this.animalAttackPenalty *= ANIMAL_ATTACK_PENALTY
     }
 
-    override fun doBeeHappy(effect: Int) {
+    override fun doBeeHappy(effect: Double) {
         this.pollination *= effect
     }
